Move TCP stream to its own module
