+++ src/stream/mod.rs
@@ -1,10 +1,9 @@
 //! Provides an abstraction over the standard `TcpStream` and `native_tls`'s `TlsStream`.
 
-use mio::Poll;
+/*
 use mio::PollOpt;
 use mio::Ready;
 use mio::Token;
-use mio::unix::EventedFd;
 use native_tls as tls;
 use net2::TcpBuilder;
 use net2::TcpStreamExt;
@@ -17,100 +16,12 @@ use std::net;
 use std::os::unix::io::{AsRawFd, RawFd};
 use std::result::Result;
 use take_mut::take;
+*/
 
-struct TcpStream<'poll> {
-    inner: net::TcpStream,
-    poll: &'poll Poll,
-    out_buf: Vec<u8>,
-}
+pub mod tcp;
+pub mod utils;
 
-impl<'poll> TcpStream<'poll> {
-    fn send(&mut self) -> io::Result<()> {
-        let to_send = self.out_buf.len();
-        match self.inner.write(&self.out_buf) {
-            Ok(bytes_sent) => {
-                self.out_buf.drain(0 .. bytes_sent);
-                let register =
-                    if bytes_sent == to_send {
-                        reregister_for_r
-                    } else {
-                        reregister_for_rw
-                    };
-                register(&self.poll, self.inner.as_raw_fd());
-                Ok(())
-            }
-            Err(err) => {
-                reregister_for_rw(&self.poll, self.inner.as_raw_fd());
-                if err.kind() == io::ErrorKind::WouldBlock {
-                    Ok(())
-                } else {
-                    Err(err)
-                }
-            }
-        }
-    }
-}
-
-impl<'poll> Drop for TcpStream<'poll> {
-    fn drop(&mut self) {
-        deregister(&self.poll, self.inner.as_raw_fd());
-    }
-}
-
-impl<'poll> Write for TcpStream<'poll> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        // TODO inefficient, sometimes it's not necessary to to copy buf to `out_buf`
-        self.out_buf.extend(buf);
-        self.send().map(|_| buf.len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        self.inner.flush()
-    }
-}
-
-impl<'poll> Read for TcpStream<'poll> {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        self.inner.read(buf)
-    }
-}
-
-#[derive(Debug)]
-pub enum TcpError {
-    IoError(io::Error),
-    CantResolveAddr,
-}
-
-impl<'poll> TcpStream<'poll> {
-    fn new(
-        poll: &'poll Poll,
-        serv_addr: &str,
-        serv_port: u16,
-    ) -> Result<TcpStream<'poll>, TcpError> {
-        let mut addr_iter = (serv_addr, serv_port)
-            .to_socket_addrs()
-            .map_err(TcpError::IoError)?;
-        let addr = addr_iter.next().ok_or(TcpError::CantResolveAddr)?;
-        let stream = {
-            match addr {
-                SocketAddr::V4(_) =>
-                    TcpBuilder::new_v4().unwrap().to_tcp_stream().unwrap(),
-                SocketAddr::V6(_) =>
-                    TcpBuilder::new_v6().unwrap().to_tcp_stream().unwrap(),
-            }
-        };
-        stream.set_nonblocking(true).unwrap();
-        // This will fail with EINPROGRESS
-        let _ = stream.connect(addr);
-        register_for_r(poll, stream.as_raw_fd());
-        Ok(TcpStream {
-            inner: stream,
-            poll: poll,
-            out_buf: Vec::with_capacity(1024),
-        })
-    }
-}
-
+/*
 enum TlsStream<'poll> {
     Handshake(tls::MidHandshakeTlsStream<TcpStream<'poll>>),
     Connected(tls::TlsStream<TcpStream<'poll>>),
@@ -205,36 +116,6 @@ impl<'poll> Stream<'poll> {
     }
 }
 
-fn register_for_r(poll: &Poll, fd: RawFd) {
-    let _ = poll.register(
-        &EventedFd(&fd),
-        Token(fd as usize),
-        Ready::readable(),
-        PollOpt::level(),
-    );
-}
-
-fn reregister_for_r(poll: &Poll, fd: RawFd) {
-    let _ = poll.reregister(
-        &EventedFd(&fd),
-        Token(fd as usize),
-        Ready::readable(),
-        PollOpt::level(),
-    );
-}
-
-fn reregister_for_rw(poll: &Poll, fd: RawFd) {
-    let _ = poll.reregister(
-        &EventedFd(&fd),
-        Token(fd as usize),
-        Ready::readable() | Ready::writable(),
-        PollOpt::level(),
-    );
-}
-
-fn deregister(poll: &Poll, fd: RawFd) {
-    let _ = poll.deregister(&EventedFd(&fd));
-}
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -304,3 +185,4 @@ mod tests {
         }
     }
 }
+*/
